requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
asttokens==2.4.1
attrs==23.2.0
beautifulsoup4==4.12.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
dataclasses-json==0.6.4
decorator==5.1.1
discord==2.3.2
discord-ext-voice-recv==0.4.2a145
discord.py @ git+https://github.com/Rapptz/discord.py.git@425edd2e10b9be3d7799c0df0cd1d43a1a34654e
distro==1.9.0
elevenlabs==1.0.2
executing==2.0.1
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.4
httpx==0.26.0
idna==3.6
ipython==8.22.2
jedi==0.19.1
jsonpatch==1.33
jsonpointer==2.4
langchain==0.1.11
langchain-community==0.0.27
langchain-core==0.1.30
langchain-openai==0.0.8
langchain-text-splitters==0.0.1
langchainhub==0.1.15
langsmith==0.1.23
marshmallow==3.21.1
matplotlib-inline==0.1.6
mediawikiapi==1.2
multidict==6.0.5
mypy-extensions==1.0.0
numpy==1.26.4
openai==1.13.3
orjson==3.9.15
packaging==23.2
parso==0.8.3
pexpect==4.9.0
prompt-toolkit==3.0.43
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
pydantic==2.6.3
pydantic_core==2.16.3
Pygments==2.17.2
PyNaCl==1.5.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
SpeechRecognition==3.10.1
SQLAlchemy==2.0.28
stack-data==0.6.3
tenacity==8.2.3
tiktoken==0.6.0
tqdm==4.66.2
traitlets==5.14.2
types-requests==2.31.0.20240311
typing-inspect==0.9.0
typing_extensions==4.10.0
upstash-redis==1.0.0
urllib3==2.2.1
wcwidth==0.2.13
websockets==12.0
wikibase-rest-api-client==0.1.3
yarl==1.9.4
